Add UserPhoneNumber model. Consolidate phone number models to use abstract model.
